Improve performance of globals
This change includes 2 improvements that will make globals faster. First
in `prepareGlobal` I'm adding an option to skip the expensive operation
of saving the entire global object so that it can later be examined.
This will be useful for when we have an exportName and therefore know
that we don't need to do the work.

The second improvement is to use an object rather than an array for
looking up globals that can be ignored.